Modify command line to use DBUS
